requirements.txt
faicons
pandas
pathlib
shiny
shinywidgets
plotly.express